Add files via upload
The final output of the cleaning process is a list that contains all the non-missing values of the 35 "data science skills related" columns. It was necessary to organize the data into a list containing named vectors because too much of the data is missing and most of the data consist of strings. Not only that, the observations are randomized (for preserving the anonymity of the respondents), meaning that each row contains column values that are not necessarily from the same individual. So, the rectangular format was not very representative of the data to begin with. Overall, for the above reasons, the list format just made the most sense.